+++ packages/react-ui/CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0-alpha.27](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.26...@selfcommunity/react-ui@0.11.0-alpha.27) (2025-02-26)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+# [0.11.0-alpha.26](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.25...@selfcommunity/react-ui@0.11.0-alpha.26) (2025-02-26)
+
+
+### Bug Fixes
+
+* category can_create_only_staff ([a9ebbf3](https://github.com/selfcommunity/community-js/commit/a9ebbf3c483bd428f63f7567227c96dcead9638f))
+
+
+
+
+
+# [0.11.0-alpha.25](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.24...@selfcommunity/react-ui@0.11.0-alpha.25) (2025-02-26)
+
+
+### Bug Fixes
+
+* live stream ([976b885](https://github.com/selfcommunity/community-js/commit/976b88550cf00cd6ec80806e4578488f4843bdff))
+
+
+
+
+
+# [0.11.0-alpha.24](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.23...@selfcommunity/react-ui@0.11.0-alpha.24) (2025-02-26)
+
+
+### Bug Fixes
+
+* live stream video processor ([e193ca4](https://github.com/selfcommunity/community-js/commit/e193ca4c9e218bb18e57eb8579cf378d67df5a62))
+
+
+
+
+
+# [0.11.0-alpha.23](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.22...@selfcommunity/react-ui@0.11.0-alpha.23) (2025-02-25)
+
+
+### Bug Fixes
+
+* live stream create button ([e2bfc4e](https://github.com/selfcommunity/community-js/commit/e2bfc4eefe3d82ec4c9c27a9ab522b19459cdeb5))
+* live stream create button ([e11a7b9](https://github.com/selfcommunity/community-js/commit/e11a7b9a034c59b581d2b203a8d6066837511f19))
+
+
+
+
+
 # [0.11.0-alpha.22](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.11.0-alpha.21...@selfcommunity/react-ui@0.11.0-alpha.22) (2025-02-25)